Refactor with Promises and internal represenations.
Major rewrite to use Promises. During this process, I've also moved to using
internal representations of anything that comes back from Telegram. This gives me
more control and moves that in the future I could add other bot types easier.

It's also organised around the message object which gets built up and then is used to
provide a response.

Started on tests.  Nothing very useful at the moment but I'm now more familiar with it
so in a better position to add more.

New feature includes being able to find the right fixture and result when home/away are mentioned.